third_party/blink/web_tests/platform/generic/fast/css/first-letter-nested-expected.txt

The three spans below should look the same.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Span

Span

Span

PASS document.getElementById('test_span_1').offsetWidth is document.getElementById('ref_span').offsetWidth
PASS document.getElementById('test_span_2').offsetWidth is document.getElementById('ref_span').offsetWidth
PASS successfullyParsed is true

TEST COMPLETE
